refactor(export): extract render-engine predicates to own module
Move the stateless GPU-effect and animated-image classifiers
(hasEnabledGpuEffect, itemHasEnabledGpuEffect,
subCompositionRenderDataHasGpuEffects, isAnimatedImage, isGifFormat) out of
client-render-engine.ts into render-engine-predicates.ts. Verbatim move,
behavior-preserving; subCompositionRenderDataHasGpuEffects is re-exported so
existing import sites and its test are unaffected.
