+++ src/features/export/utils/client-render-engine.ts
@@ -54,7 +54,6 @@ import {
 } from './canvas-masks'
 import { type ActiveTransition } from './canvas-transitions'
 import { type CachedGifFrames, gifFrameCache } from '@/features/export/deps/timeline'
-import { isGifUrl, isWebpUrl } from '@/shared/utils/media-utils'
 import { CanvasPool, TextMeasurementCache } from './canvas-pool'
 import { SharedVideoExtractorPool, type VideoFrameSource } from './shared-video-extractor'
 import { getCompositeOperation } from '@/types/blend-mode-css'
@@ -87,6 +86,12 @@ import { ScrubbingCache } from '@/features/export/deps/preview'
 import { resolveFrameRenderOptimization } from './render-path-optimizer'
 import { ReverseVideoFrameCache } from './reverse-video-frame-cache'
 import { resolveReverseConformedVideoItem } from '@/shared/utils/reverse-conform-item'
+import {
+  itemHasEnabledGpuEffect,
+  isAnimatedImage,
+  isGifFormat,
+  subCompositionRenderDataHasGpuEffects,
+} from './render-engine-predicates'
 
 // Re-export orchestration functions so existing import sites keep working
 export { renderComposition, renderAudioOnly, renderSingleFrame } from './canvas-render-orchestrator'
@@ -95,86 +100,10 @@ function getLog() {
   return createLogger('ClientRenderEngine')
 }
 
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function hasEnabledGpuEffect(effects: TimelineItem['effects']): boolean {
-  return effects?.some((e) => e.enabled && e.effect.type === 'gpu-effect') ?? false
-}
-
-function itemHasEnabledGpuEffect(
-  item: TimelineItem,
-  getPreviewEffectsOverride?: (itemId: string) => ItemEffect[] | undefined,
-): boolean {
-  const previewEffects = getPreviewEffectsOverride?.(item.id)
-  return hasEnabledGpuEffect(previewEffects ?? item.effects)
-}
-
-export function subCompositionRenderDataHasGpuEffects(
-  compositionId: string,
-  subCompRenderData: ReadonlyMap<string, SubCompRenderData>,
-  options: {
-    getCurrentItem?: <TItem extends TimelineItem>(item: TItem) => TItem
-    getPreviewEffectsOverride?: (itemId: string) => ItemEffect[] | undefined
-    visited?: Set<string>
-  } = {},
-): boolean {
-  const { getPreviewEffectsOverride } = options
-  const getCurrentItem =
-    options.getCurrentItem ?? (<TItem extends TimelineItem>(item: TItem) => item)
-  const visited = options.visited ?? new Set<string>()
-  if (visited.has(compositionId)) return false
-  visited.add(compositionId)
-
-  const subData = subCompRenderData.get(compositionId)
-  if (!subData) return false
-
-  for (const entry of subData.adjustmentLayers ?? []) {
-    if (itemHasEnabledGpuEffect(getCurrentItem(entry.layer), getPreviewEffectsOverride)) {
-      return true
-    }
-  }
-
-  for (const track of subData.sortedTracks) {
-    if (!track.visible) continue
-    for (const subItem of track.items) {
-      const currentSubItem = getCurrentItem(subItem)
-      if (itemHasEnabledGpuEffect(currentSubItem, getPreviewEffectsOverride)) return true
-      if (
-        currentSubItem.type === 'composition' &&
-        subCompositionRenderDataHasGpuEffects(currentSubItem.compositionId, subCompRenderData, {
-          getCurrentItem,
-          getPreviewEffectsOverride,
-          visited,
-        })
-      ) {
-        return true
-      }
-    }
-  }
-
-  return false
-}
-
-/**
- * Check if an image item is a potentially animated image (GIF or WebP).
- * Static WebP files will be detected during frame extraction and fall back
- * to regular image rendering.
- */
-function isAnimatedImage(item: ImageItem): boolean {
-  const label = item.label?.toLowerCase() ?? ''
-  return (
-    isGifUrl(item.src) || label.endsWith('.gif') || isWebpUrl(item.src) || label.endsWith('.webp')
-  )
-}
-
-/**
- * Check if an image item is specifically a GIF (for gifuct-js extraction).
- */
-function isGifFormat(item: ImageItem): boolean {
-  return isGifUrl(item.src) || (item.label?.toLowerCase() ?? '').endsWith('.gif')
-}
+// Predicate helpers (GPU-effect / animated-image classifiers) live in
+// `render-engine-predicates.ts`. `subCompositionRenderDataHasGpuEffects` is
+// re-exported so existing import sites (and its test) keep working.
+export { subCompositionRenderDataHasGpuEffects }
 
 // WebP frame extraction is handled by gifFrameCache.getWebpFrames() —
 // the cache service uses the ImageDecoder API and provides the same
